GASC Benchmarking Process
Caught a bug:
Calculation of the updated von misses stresses is done after the `if noslip` which updates the axial stresses

Introduced a flag 'empty_smcs' that would allow the experienced user to calculate the stresses of the worst case scenario of a series of possibilities. The casual user has it off by default.