Initial testing of core-edge integration
An initial go at using the SOL actor to inform the pedestal density boundary condition.

Whilst this does what is intended, I will need to alter this approach going forward. At the moment, the 4:1 ratio of pedestal to separatrix density is causing the pedestal density to become incredibly high, resulting in hollow current profiles (in turn crashing the equilibrium solver).

Going forward, the aim will be to flatten out the pedestal density profile which should prevent this (hopefully). This will require changing the 4:1 ratio on the fly.